Updating "college_v2.mwb"
I replaced the old "college_v2.mwb" file, which was the wrong file, with a new one. This should be the right one. 
I reverse engineered my EER diagram that has my views, tables, and all the data inside the tables. Then I saved that code to
my file and renamed it the exact same thing to replace/update the old one.