Uploading a procedure to my branch
This procedure lets the user add a new course to their database. The parameters that are inserted into the parentheses when the procedure is called have validation checks. It checks if one or more parameters are empty. If so, parameter must be filled. 

I do want to add- if course exists, don't create a duplicate. But for now, this will do.