Uploading a Function
I created a function that basically allows the user to get a department by a course. This could be used in many occasions. 
For example, say Billy wants to know where the course "Computing Principles" is. This function will take the user's input,
join the tables to access their data, filter which course you want, and return a department name.

The output for billy's question would be the Computer Science department!